+++ parsers/convert-font/README.md
@@ -1,8 +1,10 @@
-# CONVERT FONT
+# Convert Font
 
 ## Description
 
 Convert font in several formats.
+
+Learn more about how to configure Specify in the API documentation: [https://specifyapp.com/developers/cli](https://specifyapp.com/developers/cli)
 
 ## Interface 
 
@@ -19,55 +21,20 @@ interface parser {
 
 ### Options
 
-| parameter | Require    | type      | default    | description                                       |
-| --------- | ---------- | --------- | ---------- | ------------------------------------------------- |
-| `formats`    | optional   | `Array<string>`   | `["woff2", "woff", "otf", "ttf", "eot"]` | the list of formats to convert |
-| `fileNameKey`    | optional   | `name` `fontFamily` `Array<string>`   | `name` | The design token's keys that will be used to create the file name. These keys will be separated by a space to create the file name. |
-| `fileNameFormat`    | optional   | `camelCase` `kebabCase` `snakeCase`   |  | The function to normalize the file name |
+| Parameter        | Required  | Type                                  | Default                                  | Description                                       |
+| ---------------- | --------- | ------------------------------------- | ---------------------------------------- | ------------------------------------------------- |
+| `formats`        | optional     | `Array<string>`                       | `["woff2", "woff", "otf", "ttf", "eot"]` | The list of formats to convert |
+| `fileNameKey`    | optional     | `name`, `fontFamily`, `Array<string>` | `name`                                   | The design token's keys that will be used to create the file name. These keys will be separated by a space to create the file name.|
+| `fileNameFormat` | optional     | `camelCase` `kebabCase` `snakeCase`   |                                          | The function to normalize the file name |
 
-## Usage example 
-
-```json
-{
-    "name": "convert-font",
-    "options": {
-      "formats": ["woff2", "woff"],
-      "fileNameKey":  ["fontFamily", "fontWeight"],
-      "fileNameFormat": "kebabCase"
-    }
-}
-```
-
-### Result example
-
-```json
-[
-    {
-      "name": "allan-700.woff2",
-      "value": {
-        "url": "https://beta.api.specifyapp.com/...",
-        "fontFamily": "Allan",
-        "fontWeight": 700,
-        "fontPostScriptName": "Allan-Bold"
-      }
-    },
-    {
-      "name": "allan-700.woff",
-      "value": {
-        "url": "https://beta.api.specifyapp.com/...",
-        "fontFamily": "Allan",
-        "fontWeight": 700,
-        "fontPostScriptName": "Allan-Bold"
-      }
-    }
-]
-```
 
 ## Types
 
-### input
+ℹ️ **Please be aware that, depending on the order you use parsers, their input and output types have to match.**
 
-Array of object with the keys to apply kebabcase function
+### Input
+
+Array of object with at least the key `name`
 
 ```ts
 Array<{
@@ -77,10 +44,51 @@ Array<{
 }>
 ```
 
-### output
+### Output
+
 ```ts
 Array<{
   value: { fontPostScriptName: string; url: string ; [Key: string]: any };
   [Key: string]: any;
 }>
 ```
+## Usage
+
+### Config
+```json
+{
+  "name": "convert-font",
+  "options": {
+    "formats": ["woff2", "woff"],
+    "fileNameKey":  ["fontFamily", "fontWeight"],
+    "fileNameFormat": "kebabCase"
+  }
+}
+...
+```
+
+### Result
+
+```json
+[
+  {
+    "name": "allan-700.woff2",
+    "value": {
+      "url": "https://beta.api.specifyapp.com/...",
+      "fontFamily": "Allan",
+      "fontWeight": 700,
+      "fontPostScriptName": "Allan-Bold"
+    }
+  },
+  {
+    "name": "allan-700.woff",
+    "value": {
+      "url": "https://beta.api.specifyapp.com/...",
+      "fontFamily": "Allan",
+      "fontWeight": 700,
+      "fontPostScriptName": "Allan-Bold"
+    }
+  }
+]
+```
+
