+++ parsers/pick/README.md
@@ -1,8 +1,10 @@
-# PICK
+# Pick
 
 ## Description
 
 Loop on all tokens and get only keys given in params.
+
+Learn more about how to configure Specify in the API documentation: [https://specifyapp.com/developers/cli](https://specifyapp.com/developers/cli)
 
 ## Interface 
 ```ts
@@ -13,18 +15,15 @@ interface parser {
   }
 }
 ```
+### Options
 
-## Options example
-```json
-{
-  "name": "pick",
-  "options": {
-    "keys": ["name", "type"]
-  }
-}
-```
+| Parameter              | Required   | Type      | Default    | Description                                          |
+| ---------------------- | ---------- | --------- | ---------- | ---------------------------------------------------- |
+| `keys`                 | optional   | `Array`   | `["name"]` | The list of keys where the function will be applied. |
 
 ## Types
+
+ℹ️ **Please be aware that, depending on the order you use parsers, their input and output types have to match.**
 
 ### Input
 
@@ -41,43 +40,43 @@ Array of object with the keys given in options
 ```ts
 Array<{[key: string]: any}>
 ```
-
-## Before / After
-
-### Before
+## Usage
+### Config
 
 ```json
-[
-    {
-        "type": "color",
-        "value": {
-            "a": 0.96,
-            "b": 20,
-            "g": 227,
-            "r": 122
-        },
-        "name": "Generic Borders generate"
-    }
-]
-```
-
-### Options used on parser
-
-```
 {
   "name": "pick",
   "options": {
-    "keys": ["name"]
+    "keys": ["name", "type"]
   }
 }
+...
 ```
+### Before/After
 
-### Result
+#### Input
+
+```js
+[
+  {
+    "type": "color",
+    "value": {
+      "a": 0.96,
+      "b": 20,
+      "g": 227,
+      "r": 122
+    },
+    "name": "Brand / Primary Color"
+  }
+]
+```
+#### Output
 
 ```json
 [
   {
-    "name": "Generic Borders generate"
+    "name": "Brand / Primary Color",
+    "type": "color"
   }
 ]
 ```
