+++ parsers/suffix-by/README.md
@@ -1,8 +1,10 @@
-# SUFFIX BY
+# Suffix By
 
 ## Description
 
-Allows to concatenate two strings
+Allows to concatenate two strings.
+
+Learn more about how to configure Specify in the API documentation: [https://specifyapp.com/developers/cli](https://specifyapp.com/developers/cli)
 
 ## Interface 
 
@@ -18,28 +20,17 @@ interface parser {
 ```
 
 ### Options
-| parameter | Require    | type             | default    | description                                       |
-| --------- | ---------- | ---------------- | ---------- | ------------------------------------------------- |
-| `key`     | optional   | `string`         | "name"     | the key of the value will be suffixed             |
-| `suffix`  | required   | `string`         |            | the string used as content to suffix              |
-| `types`   | optional   | `Array<string>`  |            | the types where the function will be applied        |
-
-## Usage example 
-
-```json
-{
-    "name": "suffix-by",
-    "options": {
-      "types": ["vector"],
-      "suffix": ".svg",
-      "key": "name"
-    }
-}
-```
+| Parameter | Required | Type             | Default | Description                                  |
+| --------- | -------- | ---------------- | ------- | -------------------------------------------- |
+| `key`     | optional | `string`         | "name"  | The key of the value will be suffixed        |
+| `suffix`  | required | `string`         |         | The string used as content to suffix         |
+| `types`   | optional | `Array<string>`  |         | The types where the function will be applied |
 
 ## Types
 
-### input
+ℹ️ **Please be aware that, depending on the order you use parsers, their input and output types have to match.**
+
+### Input
 
 Array of object with the keys to apply the suffix
 
@@ -47,7 +38,23 @@ Array of object with the keys to apply the suffix
 Array<{[key: string]: any}>
 ```
 
-### output
-```
+### Output
+
+```ts
 Array<{[key: string]: any}>
+```
+
+## Usage
+### Config
+
+```json
+{
+  "name": "suffix-by",
+  "options": {
+    "types": ["vector"],
+    "suffix": ".svg",
+    "key": "name"
+  }
+}
+...
 ```
