+++ parsers/svgo/svgo.parser.ts
