+++ parsers/to-dsp/README.md
@@ -11,43 +11,28 @@ interface x {
   name: 'to-dsp';
   options?: Partial<{
     settings: Partial<{
-      name?: string; // The name of the DSP (default: 'Generated by Specify')
-      buildStatusLabel?: string; // Build status label (examples : 'dev' | 'released')
-      packageVersion?: string; // Version of the DSP package, not the DSP spec
-      snippetTriggerPrefix?: string; // Trigger for snippets on components (examples: 'sp-' | 'ex-")
+      name?: string;
+      buildStatusLabel?: string;
+      packageVersion?: string;
+      snippetTriggerPrefix?: string;
     }>;
-    createAssets: boolean; // Defines if you want the assets to be created or not (default: true)
+    createAssets: boolean;
   }>;
 }
 ```
 
-## Options examples
-
-```json
-{
-  "name": "to-dsp",
-  "options": {
-    "settings": {
-      "name": "My awesome DSP"
-    }
-  }
-}
-```
-
-```json
-{
-  "name": "to-dsp",
-  "options": {
-    "settings": {
-      "name": "My awesome DSP",
-      "buildStatusLabel": "beta"
-    },
-    "createAssets": false
-  }
-}
-```
+### Options
+| Parameter                       | Required  | Type     | Default                | Description                                                   |
+| ------------------------------- | --------- | -------- | ---------------------- | ------------------------------------------------------------- |
+| `settings.name`                 | optional     | `string` | 'Generated by Specify' | The name of the DSP                                           |
+| `settings.buildStatusLabel`     | optional     | `string` |                        | Build status label (examples: 'dev' or 'released')            |
+| `settings.packageVersion`       | optional     | `string` |                        | Version of the DSP package, not the DSP spec                  |
+| `settings.snippetTriggerPrefix` | optional     | `string` |                        | Trigger for snippets on components (examples: 'sp-' or 'ex-") |
+| `createAssets`                  | true      | `true`   | `true`                 | Defines if you want the assets to be created or not           |
 
 ## Types
+
+ℹ️ **Please be aware that, depending on the order you use parsers, their input and output types have to match.**
 
 ### Input
 
@@ -76,9 +61,27 @@ Array<{
 }>
 ```
 
-## Before / After
+## Usage
+### Config
 
-### Before
+```json
+{
+  "name": "to-dsp",
+  "options": {
+    "settings": {
+      "name": "My awesome DSP",
+      "buildStatusLabel": "test",
+      "packageVersion": "0.0.1",
+      "snippetTriggerPrefix": "ex-"
+    },
+    "createAssets": false
+  }
+}
+...
+```
+### Before/After
+
+#### Input
 
 ```json
 [
@@ -94,25 +97,7 @@ Array<{
   }
 ]
 ```
-
-### Options used on parser
-
-```json
-{
-  "name": "to-dsp",
-  "options": {
-    "settings": {
-      "name": "My awesome DSP",
-      "buildStatusLabel": "test",
-      "packageVersion": "0.0.1",
-      "snippetTriggerPrefix": "ex-"
-    },
-    "createAssets": false
-  }
-}
-```
-
-### Result
+#### Output
 
 ```json
 [
@@ -128,9 +113,14 @@ Array<{
         "        {\n" +
         "            \"class\": \"token\",\n" +
         "            \"type\": \"color\",\n" +
-        "            \"id\": \"96218bbe-0948-4822-a03d-1aa9ea2abe4d\",\n" +
-        "            \"name\": \"heuristic cross-platform quantify\",\n" +
-        "            \"value\": \"#330f6399\",\n" +
+        "            \"name\": \"Primary color\",\n" +
+        "            \"value\": \n" +
+                      "{\n" +
+                      "    \"a\": \"0.96\",\n" +
+                      "    \"b\": \"20\",\n" +
+                      "    \"g\": \"227\",\n" +
+                      "    \"r\": \"122\",\n" +
+                      "}," +
         "            \"tags\": [\n" +
         "                \"specify\",\n" +
         "                \"color\"\n" +
